requirements.txt
APScheduler==3.10.4
aiogram==2.24
aioschedule==0.5.2
asyncpg==0.27.0
python-dotenv==1.0.0